42filechecker

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: cvernius <cvernius@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/16 21:19:10 by cvernius          #+#    #+#             */
-/*   Updated: 2019/05/10 15:45:45 by cvernius         ###   ########.fr       */
+/*   Updated: 2019/05/11 16:58:34 by cvernius         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_atoi(const char *str)
 {
-	int			total;
+	long		total;
 	size_t		i;
 	int			flag;
 
@@ -29,8 +29,8 @@ int		ft_atoi(const char *str)
 			flag = -1;
 		i++;
 	}
-	while (str[i] >= '0' && str[i] <= '9' && str[i] != '\0' && total <= INT_MAX
-														&& total >= INT_MIN)
+	while (str[i] >= '0' && str[i] <= '9' && str[i] != '\0' && total <= LONG_MAX
+														&& total >= LONG_MIN)
 	{
 		total = total * 10 + (str[i] - 48);
 		i++;
